Moved code into separate module
