+++ tests/conftest.py
@@ -5,9 +5,6 @@ from unittest.mock import Mock
 
 import pytest
 
-from OTAnalytics.application.datastore import FlowParser
-from OTAnalytics.application.parser.cli_parser import CliArguments
-from OTAnalytics.application.run_configuration import RunConfiguration
 from OTAnalytics.domain.geometry import Coordinate
 from OTAnalytics.domain.section import LineSection, Section, SectionId
 from OTAnalytics.domain.track import Track, TrackId
@@ -320,46 +317,3 @@ def pandas_track_segment_dataset_builder(
     track_segment_dataset_builder_provider: TrackSegmentDatasetBuilderProvider,
 ) -> TrackSegmentDatasetBuilder:
     return track_segment_dataset_builder_provider.provide(PANDAS)
-
-
-NUM_PROCESSES = 1
-
-
-def create_run_config(
-    track_files: list[str],
-    otflow_file: str,
-    save_dir: str,
-    event_formats: list[str],
-    flow_parser: FlowParser,
-    start_cli: bool = True,
-    debug: bool = False,
-    logfile_overwrite: bool = True,
-    track_export: bool = False,
-    num_processes: int = NUM_PROCESSES,
-    count_intervals: list[int] | None = None,
-    config_file: str | None = None,
-    save_name: str | None = None,
-    save_suffix: str | None = None,
-    log_file: str | None = None,
-    include_classes: list[str] | None = None,
-    exclude_classes: list[str] | None = None,
-) -> RunConfiguration:
-    cli_args = CliArguments(
-        start_cli=start_cli,
-        debug=debug,
-        logfile_overwrite=logfile_overwrite,
-        track_export=track_export,
-        track_files=track_files,
-        otflow_file=otflow_file,
-        save_dir=save_dir,
-        save_name=save_name,
-        count_intervals=count_intervals,
-        event_formats=event_formats,
-        num_processes=num_processes,
-        config_file=config_file,
-        save_suffix=save_suffix,
-        log_file=log_file,
-        include_classes=include_classes,
-        exclude_classes=exclude_classes,
-    )
-    return RunConfiguration(flow_parser, cli_args)
